Fix a bunch of lint issues.
R=chrisha@chromium.org

Review URL: https://codereview.appspot.com/12649043

git-svn-id: http://sawbuck.googlecode.com/svn/trunk@1720 15e8cca8-e42c-11de-a347-f34a4f72eb7d

+++ syzygy/agent/asan/asan_rtl_impl.cc
@@ -658,7 +658,7 @@ SIZE_T WINAPI asan_HeapSize(HANDLE heap,
 
   HeapProxy* proxy = HeapProxy::FromHandle(heap);
   if (!proxy)
-    return -1;
+    return static_cast<SIZE_T>(-1);
 
   return proxy->Size(flags, mem);
 }
